backend/users-service/requirements.txt
Django>=6.0.2
pika>=1.3.0
psycopg2-binary>=2.9
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.0
python-dotenv>=1.0.1
pytest>=7.0
pytest-django>=4.5
